backend/Noorgpt/requirements.txt
flask
requests
gunicorn
flask-cors
flask-sqlalchemy
psycopg2-binary
python-dotenv
Flask-Mail
